Contracts: Move all utils and helpers to utils dir
